Move split_components
